+++ src/popup/warnings/form.jsx
@@ -1,21 +1,39 @@
 import './form.scss'
 import React, { useState } from 'react'
+import ColorField from '../color-field'
 import { warningStyles } from './friendly-names'
 
 export default function WarningForm ({ onSave, value, disabled = false }) {
-  const [pattern, setPattern] = useState(
-    value?.pattern ?? ''
-  )
+  const [pattern, setPattern] = useState(value?.pattern ?? '')
   const [warningStyle, setWarningStyle] = useState(
     value?.warningStyle ?? Object.keys(warningStyles)[0]
   )
+  const [text, setText] = useState(value?.text ?? 'Warning! This is Production!')
+  const [borderColor, setBorderColor] = useState(value?.borderColor ?? 'FF0000')
+  const [backgroundColor, setBackgroundColor] = useState(value?.backgroundColor ?? 'FF0000')
+  const [textColor, setTextColor] = useState(value?.textColor ?? 'FFFFFF')
 
   function handleSubmit (event) {
     event.preventDefault()
-    onSave({
+    const payload = {
       pattern,
       warningStyle
-    })
+    }
+
+    switch (warningStyle) {
+      case 'bottomBanner':
+      case 'topBanner':
+        payload.text = text
+        payload.backgroundColor = backgroundColor
+        payload.textColor = textColor
+        break
+
+      case 'border':
+        payload.borderColor = borderColor
+        break
+    }
+
+    onSave(payload)
   }
 
   return (
@@ -23,52 +41,76 @@ export default function WarningForm ({ onSave, value, disabled = false }) {
       className='warning-form'
       onSubmit={handleSubmit}
     >
-      <label className='field'>
-        <span>URL Pattern:</span>
-        <input
-          type='text'
-          required
-          value={pattern}
-          onChange={e => setPattern(e.target.value)}
-          disabled={disabled}
-        />
-      </label>
+      <label>URL Regex:</label>
+      <input
+        type='text'
+        required
+        value={pattern}
+        onChange={e => setPattern(e.target.value)}
+        disabled={disabled}
+      />
 
-      <FieldHelp>
-        A regular expression matched against a tab's URL.
-        If there's a match, the warning is displayed.
-      </FieldHelp>
+      <label>Style:</label>
+      <select
+        required
+        value={warningStyle}
+        onChange={e => setWarningStyle(e.target.value)}
+        disabled={disabled}
+      >
+        {Object.entries(warningStyles)
+          .map(([key, name]) => (
+            <option key={key} value={key}>{name}</option>
+          ))}
+      </select>
 
-      <label className='field'>
-        <span>Style:</span>
-        <select
-          required
-          value={warningStyle}
-          onChange={e => setWarningStyle(e.target.value)}
-          disabled={disabled}
-        >
-          {Object.entries(warningStyles)
-            .map(([key, name]) => (
-              <option key={key} value={key}>{name}</option>
-            ))}
-        </select>
-      </label>
+      {warningStyle === 'border'
+        ? (
+          <>
+            <label>Border Color:</label>
+            <ColorField
+              value={borderColor}
+              onChange={setBorderColor}
+              disabled={disabled}
+            />
+          </>)
+        : null}
 
-      <FieldHelp>
-        Controls what kind of warning to display.
-      </FieldHelp>
+      {['topBanner', 'bottomBanner'].includes(warningStyle)
+        ? (
+          <>
+            <label>Message:</label>
+            <input
+              type='text'
+              required
+              value={text}
+              onChange={e => setText(e.target.value)}
+              disabled={disabled}
+            />
+
+            <label>Text Color:</label>
+            <ColorField
+              value={textColor}
+              onChange={setTextColor}
+              disabled={disabled}
+            />
+
+            <label>Background Color:</label>
+            <ColorField
+              value={backgroundColor}
+              onChange={setBackgroundColor}
+              disabled={disabled}
+            />
+          </>)
+        : null}
 
       <div className='controls'>
-        <button type='submit'>Save</button>
+        <button
+          type='submit'
+          disabled={disabled}
+        >
+          Save
+        </button>
       </div>
     </form>
-  )
-}
-
-function FieldHelp ({ children }) {
-  return (
-    <div className='field-help'>
-      <div className='text'>{children}</div>
-    </div>
   )
 }